requirements.txt
dj-database-url==0.5.0
Django==3.0.5
gunicorn==20.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
pytest==5.4.1
pytest-django==3.9.0
selenium==3.141.0
whitenoise==5.0.1
psycopg2==2.8.5
sentry-sdk==0.14.3
requests==2.23.0
drf-yasg==1.17.1